backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.52.0
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10